Use QVector<float> in level of detail APIs
For consistency with rest of API and to avoid ambiguity inherent with
qreal.

Task-number: QTBUG-58888
Task-number: QTBUG-58889

+++ src/render/frontend/qlevelofdetail.h
@@ -59,7 +59,7 @@ class QT3DRENDERSHARED_EXPORT QLevelOfDetail : public Qt3DCore::QComponent
     Q_PROPERTY(Qt3DRender::QCamera *camera READ camera WRITE setCamera NOTIFY cameraChanged)
     Q_PROPERTY(int currentIndex READ currentIndex WRITE setCurrentIndex NOTIFY currentIndexChanged)
     Q_PROPERTY(ThresholdType thresholdType READ thresholdType WRITE setThresholdType NOTIFY thresholdTypeChanged)
-    Q_PROPERTY(QVector<qreal> thresholds READ thresholds WRITE setThresholds NOTIFY thresholdsChanged)
+    Q_PROPERTY(QVector<float> thresholds READ thresholds WRITE setThresholds NOTIFY thresholdsChanged)
     Q_PROPERTY(Qt3DRender::QBoundingSphere *volumeOverride READ volumeOverride WRITE setVolumeOverride NOTIFY volumeOverrideChanged)
 
 public:
@@ -75,21 +75,21 @@ public:
     QCamera *camera() const;
     int currentIndex() const;
     ThresholdType thresholdType() const;
-    QVector<qreal> thresholds() const;
+    QVector<float> thresholds() const;
     QBoundingSphere *volumeOverride() const;
 
 public Q_SLOTS:
     void setCamera(QCamera *camera);
     void setCurrentIndex(int currentIndex);
     void setThresholdType(ThresholdType thresholdType);
-    void setThresholds(QVector<qreal> thresholds);
+    void setThresholds(QVector<float> thresholds);
     void setVolumeOverride(QBoundingSphere *volumeOverride);
 
 Q_SIGNALS:
     void cameraChanged(QCamera *camera);
     void currentIndexChanged(int currentIndex);
     void thresholdTypeChanged(ThresholdType thresholdType);
-    void thresholdsChanged(QVector<qreal> thresholds);
+    void thresholdsChanged(QVector<float> thresholds);
     void volumeOverrideChanged(QBoundingSphere *volumeOverride);
 
 protected:
